+++ dataweaver/apps/web/src/components/elements/card/text.tsx
@@ -1,10 +1,6 @@
-import type { DOMNode } from 'html-react-parser';
-import parse, { domToReact, Element } from 'html-react-parser';
-import Link from 'next/link';
-import { useMemo } from 'react';
 import type { CardState } from '~/components/elements/card/base';
 import { Skeleton } from '~/components/elements/skeleton';
-import { validate_href } from '~/functions/validate_href';
+import { HtmlParsed } from '~/components/primitives/html_parsed';
 import s from './text.module.scss';
 
 export interface CardTextProps extends Pick<CardState, 'isLoading'> {
@@ -21,74 +17,15 @@ export const CardText = ({
   isLoading,
   onAction,
 }: CardTextProps) => {
-  const parsedHtml = useMemo(
-    () =>
-      body
-        ? parse(body, {
-            replace: (domNode) => {
-              if (domNode instanceof Element) {
-                switch (domNode.name) {
-                  case 'script':
-                  case 'style':
-                  case 'iframe':
-                    return null;
-                  case 'a': {
-                    const validLink = validate_href(domNode.attribs.href);
-                    if (!validLink) {
-                      return <>{domToReact(domNode.children as DOMNode[])}</>;
-                    }
-
-                    const { href, target, rel } = validLink;
-
-                    // Internal action links — fire callback instead of navigating.
-                    if (href.startsWith('#') && onAction) {
-                      return (
-                        <a
-                          href={href}
-                          onPointerDown={(e) => e.stopPropagation()}
-                          onClick={(e) => {
-                            e.preventDefault();
-                            onAction(href);
-                          }}
-                        >
-                          {domToReact(domNode.children as DOMNode[])}
-                        </a>
-                      );
-                    }
-
-                    return (
-                      <Link href={href} target={target} rel={rel}>
-                        {domToReact(domNode.children as DOMNode[])}
-                      </Link>
-                    );
-                  }
-                  case 'hr':
-                    return (
-                      <hr
-                        key={
-                          'startIndex' in domNode
-                            ? domNode.startIndex
-                            : undefined
-                        }
-                      />
-                    );
-                }
-              }
-            },
-          })
-        : null,
-    [body, onAction],
-  );
-
   return (
     <>
       {title && <h2 className={s.title}>{title}</h2>}
 
       {isLoading ? (
         <Skeleton />
-      ) : (
-        parsedHtml && <div className={s.body}>{parsedHtml}</div>
-      )}
+      ) : body ? (
+        <HtmlParsed html={body} onAction={onAction} className={s.body} />
+      ) : null}
     </>
   );
 };
