--- dataweaver/packages/tokens/src/color_swatches.html
@@ -1,75 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<title>Color Swatches — @package/tokens</title>
-<style>
-  * { margin: 0; padding: 0; box-sizing: border-box; }
-  body { font-family: system-ui, sans-serif; padding: 32px; background: #f5f5f5; }
-  h1 { margin-bottom: 24px; font-size: 20px; color: #333; }
-  .grid { display: grid; grid-template-columns: repeat(auto-fill, minmax(220px, 1fr)); gap: 16px; }
-  .swatch { background: #fff; border-radius: 10px; overflow: hidden; box-shadow: 0 1px 3px rgba(0,0,0,.12); }
-  .swatch-color { height: 80px; position: relative; }
-  .swatch-color span { position: absolute; bottom: 6px; right: 8px; font-size: 11px; padding: 2px 6px; border-radius: 4px; }
-  .swatch-info { padding: 10px 12px; }
-  .swatch-name { font-size: 13px; font-weight: 600; color: #222; word-break: break-all; }
-  .swatch-rgb  { font-size: 12px; color: #666; margin-top: 2px; font-family: monospace; }
-</style>
-</head>
-<body>
-<h1>Design Token Color Swatches</h1>
-<div class="grid" id="grid"></div>
-<script>
-const colors = {
-  "surface-base": [247, 252, 255],
-  "surface-decorator": [227, 231, 237],
-  "surface-content": [0, 0, 0],
-  "control-surface": [247, 252, 255],
-  "control-surface-raised": [255, 255, 255],
-  "control-surface-hover": [227, 231, 237],
-  "control-accent": [11, 87, 208],
-  "control-accent-content": [255, 255, 255],
-  "control-content": [32, 33, 36],
-  "control-content-muted": [68, 71, 70],
-  "button-base": [194, 231, 255],
-  "button-base-hover": [173, 214, 240],
-  "button-content": [0, 0, 0],
-  "button-content-hover": [0, 0, 0],
-  "card-base": [255, 255, 255],
-  "card-base-subtle": [248, 249, 252],
-  "card-base-selected": [11, 87, 208],
-  "card-content": [32, 33, 36],
-  "card-content-muted": [68, 71, 70],
-  "card-content-subtle": [95, 99, 104],
-  "card-content-dimmed": [196, 199, 197],
-  "card-skeleton": [242, 242, 242],
-  "toaster-surface": [255, 255, 255],
-  "toaster-surface-hover": [245, 245, 245],
-  "toaster-content": [32, 33, 36],
-  "toaster-content-muted": [68, 71, 70],
-  "shadow": [0, 0, 0]
-};
-
-const grid = document.getElementById('grid');
-
-for (const [name, [r, g, b]] of Object.entries(colors)) {
-  const luminance = (0.299 * r + 0.587 * g + 0.114 * b) / 255;
-  const textColor = luminance > 0.5 ? '#000' : '#fff';
-
-  const el = document.createElement('div');
-  el.className = 'swatch';
-  el.innerHTML = `
-    <div class="swatch-color" style="background: rgb(${r}, ${g}, ${b})">
-      <span style="color:${textColor}; background: ${textColor === '#000' ? 'rgba(0,0,0,.08)' : 'rgba(255,255,255,.15)'}">
-        #${r.toString(16).padStart(2,'0')}${g.toString(16).padStart(2,'0')}${b.toString(16).padStart(2,'0')}
-      </span>
-    </div>
-    <div class="swatch-info">
-      <div class="swatch-name">${name}</div>
-      <div class="swatch-rgb">rgb(${r}, ${g}, ${b})</div>
-    </div>`;
-  grid.appendChild(el);
-}
-</script>
-</body>
-</html>
